Submit moves to the game in the base MinesweeperAI class
The subclasses ultimately just need to decide what information
they should derive and track after each turn and which square
should be revealed in the next move. The base class can then
submit that square to reveal to the game since the step of
actually making the move will always happen the same way.